fix: all mcp clients used by prototype runtimes are now closed after the agent exits
